refactor: charger_card.dart component 로 이동
